+++ changelog.md
@@ -30,6 +30,7 @@
 * LLT-3628: Enable IPv6 and add tests for it (unit and nat-lab).
 * LLT-4042: IPv6 firewall nat-lab tests.
 * LLT-3914: Add apple tvOS support
+* LLT-4013: Add conntrack rules for icmp
 
 <br>